src/backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
python-dotenv==1.0.1
supabase==2.5.0
google-generativeai==0.7.2
pdfplumber==0.11.1